basic contact form working with desktop
