--- src/components/shared/SuccessModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SuccessModal';
